requirements.txt
streamlit==1.25.0
pypdf==3.8.1 
gtts==2.3.2
groq
langchain
openai
